fastapi_pt/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
prometheus-fastapi-instrumentator==6.1.0
tritonclient[all]==2.41.0
torch==2.5.1+cpu
torchvision==0.20.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
